Fix error in genClusteredChain
If the background array was of type boolean (e.g. if the entries are
Bernoulli), then the diagonal blocks also got cast to Boolean.  So if
the diagonal blocks were drawn from a uniform distributions, all entries
would become 1.  Hence, after normalization, each row of each diagonal block
would have all entries the same.

To fix, cast the background array to double.

+++ genClusteredChain.m
@@ -14,7 +14,7 @@ function P = genClusteredChain(sizes, intra, inter)
 
   n = sum(sizes);
   k = length(sizes);
-  P = inter(n, n);
+  P = double(inter(n, n));
   m = 1;
   for s = sizes
     M = m + s - 1;
